requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
click==8.1.8
discord.py==2.5.2
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
typing_extensions==4.13.2
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
